Guide/Constraint Programming/requirements.txt

flake8==3.8.3
Flask==1.1.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
Werkzeug==1.0.1
certifi==2020.6.20
chardet==3.0.4
gunicorn==20.0.4
requests==0.11.1
python-constraint
